backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.2
pydantic==2.9.0
pydantic-settings==2.5.2
python-jose[cryptography]==3.4.0
httpx==0.27.2
celery==5.4.0
redis==5.1.0
boto3==1.35.0
python-multipart==0.0.20
zep-cloud==2.4.0
openai==1.51.0
tiktoken==0.8.0
pandas==2.2.3
numpy==2.1.1
jinja2==3.1.6
weasyprint==62.3
xhtml2pdf==0.2.17
tenacity==9.0.0
structlog==24.4.0
pytest==8.3.3
pytest-asyncio==0.24.0
slowapi==0.1.9
limits==3.13.0
